drivers: ethernet: Add doxygen to net_eth_mac_load and move to public API
Add missing doxygen brief to the net_eth_mac_load helper function.
