+++ drivers/ethernet/eth.h
@@ -43,38 +43,4 @@ static inline void gen_random_mac(uint8_t *mac_addr, uint8_t b0, uint8_t b1, uin
 	sys_rand_get(&mac_addr[3], 3U);
 }
 
-static inline int net_eth_mac_load(const struct net_eth_mac_config *cfg, uint8_t *mac_addr)
-{
-	if (cfg == NULL || cfg->type == NET_ETH_MAC_DEFAULT) {
-		return -ENODATA;
-	}
-
-	/* Copy the static part */
-	memcpy(mac_addr, cfg->addr, cfg->addr_len);
-
-	if (cfg->type == NET_ETH_MAC_RANDOM) {
-		sys_rand_get(&mac_addr[cfg->addr_len], NET_ETH_ADDR_LEN - cfg->addr_len);
-
-		/* Clear group bit, multicast (I/G) */
-		mac_addr[0] &= ~0x01;
-		/* Set MAC address locally administered, unicast (LAA) */
-		mac_addr[0] |= 0x02;
-
-		return 0;
-	}
-
-#if defined(CONFIG_NVMEM)
-	if (cfg->type == NET_ETH_MAC_NVMEM) {
-		return nvmem_cell_read(&cfg->cell, &mac_addr[cfg->addr_len], 0,
-				       NET_ETH_ADDR_LEN - cfg->addr_len);
-	}
-#endif
-
-	if (cfg->type == NET_ETH_MAC_STATIC) {
-		return 0;
-	}
-
-	return -ENODATA;
-}
-
 #endif /* ZEPHYR_DRIVERS_ETHERNET_ETH_H_ */
